+++ src/nnfusion/core/program/program.cpp
@@ -1,11 +1,11 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-// Metagraph IR, which is to guide the codegen procedcure.
-// This IR is based on ONNIX::ir's interface, but
+// Metagraph program, which is to guide the codegen procedcure.
+// This program is based on ONNIX::ir's interface, but
 // Instructions has attribute, namespace(todo), and tag
 
-#include "IR.hpp"
+#include "program.hpp"
 
 using namespace nnfusion::program;
