+++ src/nnfusion/engine/interpreter.hpp
@@ -5,7 +5,7 @@
 
 #include "nnfusion/common/common.hpp"
 #include "nnfusion/common/descriptor/tensor.hpp"
-#include "nnfusion/core/IR/IR.hpp"
+#include "nnfusion/core/program/program.hpp"
 #include "nnfusion/engine/memory_allocator.hpp"
 #include "nnfusion/engine/pass/graph/graph_pass.hpp"
 #include "op.hpp"
