+++ test/nnfusion/core/tagable.cpp
@@ -12,8 +12,8 @@
 
 #include "../test_util/common.hpp"
 #include "gtest/gtest.h"
-#include "nnfusion/core/IR/IR.hpp"
 #include "nnfusion/core/graph/gnode.hpp"
+#include "nnfusion/core/program/program.hpp"
 #include "nnfusion/engine/profiler/profiler.hpp"
 
 TEST(nnfusion_core_ir, tagable)
